+++ phoss-smp-webapp/src/test/java/com/helger/phoss/smp/domain/servicegroup/ISMPServiceGroupManagerFuncTest.java
@@ -29,9 +29,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestRule;
 
-import com.helger.peppol.identifier.IParticipantIdentifier;
-import com.helger.peppol.identifier.factory.IIdentifierFactory;
-import com.helger.peppol.identifier.peppol.PeppolIdentifierHelper;
+import com.helger.peppolid.IParticipantIdentifier;
+import com.helger.peppolid.factory.IIdentifierFactory;
+import com.helger.peppolid.peppol.PeppolIdentifierHelper;
 import com.helger.phoss.smp.domain.SMPMetaManager;
 import com.helger.phoss.smp.domain.user.ISMPUserManager;
 import com.helger.phoss.smp.exception.SMPServerException;
@@ -80,7 +80,7 @@ public final class ISMPServiceGroupManagerFuncTest
     try
     {
       // Check empty state
-      final int nCount = aSGMgr.getSMPServiceGroupCount ();
+      final long nCount = aSGMgr.getSMPServiceGroupCount ();
       assertEquals (nCount, aSGMgr.getAllSMPServiceGroups ().size ());
       assertFalse (aSGMgr.containsSMPServiceGroupWithID (aPI1));
       assertFalse (aSGMgr.containsSMPServiceGroupWithID (aPI2));
